include useRef hook on Island compomponent
useRef is a React Hook that provides a way to create a mutable object that persists throughout the lifecycle of a component.

It's often used to obtain references to DOM elements or, in this case, 3D objects in a React Three Fiber (R3F) environment.

In your example, const islandRef = useRef(); is creating a ref object called islandRef.

This ref will be attached to the 3D model rendered by the Island component, specifically to the <a.group> element.

 The ref attribute is then set to this islandRef, which allows you to access and interact with the underlying 3D object.